+++ src/Spectra.jl
@@ -1,5 +1,5 @@
 #############################################################################
-#Copyright (c) 2016 Charles Le Losq
+#Copyright (c) 2016-2017 Charles Le Losq
 #
 #The MIT License (MIT)
 #
@@ -11,7 +11,7 @@
 #
 #############################################################################
 
-VERSION >= v"0.5.0" && __precompile__()
+VERSION >= v"0.6.0" && __precompile__()
 
 module Spectra
 
@@ -23,6 +23,7 @@ using Dierckx
 using Ipopt
 using JuMP
 using NMF
+using Polynomials
 
 # For PyCall modules
 const preprocessing = PyNULL()
@@ -66,7 +67,7 @@ export trapz, bandarea
 export peak_diffusion, model, IRdataprep
 
 #From functions.jl
-export poly, polyfit, gaussiennes, lorentziennes, pseudovoigts, pearson7, normal_dist, xshift_inversion, xshift_direct,xshift_correction, SavitzkyGolayFilter, smooth
+export poly, polyfit, gaussiennes, lorentziennes, pseudovoigts, pearson7, normal_dist, xshift_inversion, xshift_direct,xshift_correction, SavitzkyGolayFilter, smooth, flipsp, resample
 
 #From baseline.jl
 export baseline, whitsmdd