+++ src/tlcorrection.jl
@@ -1,5 +1,5 @@
 #############################################################################
-#Copyright (c) 2016 Charles Le Losq
+#Copyright (c) 2016-2017 Charles Le Losq
 #
 #The MIT License (MIT)
 #
@@ -107,7 +107,7 @@ function tlcorrection(data::Array{Float64},temp::Float64,wave::Float64;correctio
 	elseif correction =="hehlen"
 		# this uses the formula reported in Hehlen et al. 2010
     	frequency = 1./(nu0.^3.*density) # frequency + density correction; M/KG
-    	boltzman = 1 - exp(-h.*c.*nu./(k.*T)) # dimensionless
+    	boltzman = 1 - exp.(-h.*c.*nu./(k.*T)) # dimensionless
     	ycorr = y.*frequency.*boltzman; # correction
 	else
 		error("Not implemented, choose between long, galeener or hehlen.")